// resources/js/api/generated/models/getApiV1ImageUpdates200UpdatesItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1ImageUpdates200UpdatesItem = {
  check_error: string;
  container_name: string;
  created_at: string;
  current_image_name: string;
  current_repo_digest: string;
  /** @minimum 0 */
  id: number;
  last_checked_at: string;
  latest_repo_digest: string;
  /** @minimum 0 */
  server_id: number;
  stack_name: string;
  update_available: boolean;
  updated_at: string;
};